feat: 스프린트 미션4 (05: File 입출력 관련 개선)
- BinaryContent 엔티티 필드 개선
- DTO 개선(BinaryContentCreateRequest, MessageCreateRequest, MessageUpdateRequest, UserCreateRequest, UserUpdateRequest)
- BinaryContentRepository에서 findBinaryContentByContentPath() 메서드 삭제
- UserService, MessageService 인터페이스에서 create, update 메서드 파라미터에 MultipartFile 추가
- BasicUserService, BasicMessageService, BinaryContentRepository에서 create, update 메서드에서 BinaryContent 생성 로직 개선
- UserController, MessageController에서 create, update 메서드 파라미터에 requestDTO와 MultipartFile 추가 -> Postman에서 form-date로 보내도록 개선
- FileException 예외 클래스 추가
- GlobalExceptionHandler에서 FileException 예외 처리
- DiscodeitApplication에서 테스트 코드 주석 처리

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicMessageService.java
@@ -5,6 +5,7 @@ import com.sprint.mission.discodeit.dto.request.MessageUpdateRequest;
 import com.sprint.mission.discodeit.dto.response.MessageUpdateResponse;
 import com.sprint.mission.discodeit.entity.BinaryContent;
 import com.sprint.mission.discodeit.entity.Message;
+import com.sprint.mission.discodeit.exception.FileException;
 import com.sprint.mission.discodeit.exception.ObjectNotFoundException;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.repository.ChannelRepository;
@@ -14,7 +15,9 @@ import com.sprint.mission.discodeit.service.MessageService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
 
+import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.UUID;
@@ -32,7 +35,7 @@ public class BasicMessageService implements MessageService {
 
     //interface
     @Override
-    public Message createMessage(MessageCreateRequest request) {
+    public Message createMessage(MessageCreateRequest request, List<MultipartFile> files) {
         //입력값 검증 처리하겠습니다
 //        validateString(request.content());
 //        validateUUID(request.authorId());
@@ -42,15 +45,25 @@ public class BasicMessageService implements MessageService {
         validateUserExists(request.authorId());
         validateChannelExists(request.channelId());
 
-        //첨부파일 추가 작업
         List<UUID> binaryContentIdList = new ArrayList<>();
-        if (request.attachmentPathList() != null) {
-            for (String attachmentPath : request.attachmentPathList()) {
-                if (attachmentPath != null && !attachmentPath.isBlank()) {
-                    //binaryContent 생성
-                    BinaryContent binaryContent = new BinaryContent(attachmentPath);
-                    binaryContentRepository.createBinaryContent(binaryContent);
-                    binaryContentIdList.add(binaryContent.getId());
+        //첨부파일 존재 시
+        if (files != null) {
+            for (MultipartFile file : files) {
+                if (file != null && !file.isEmpty()) {
+                    try {
+                        //binaryContent 생성
+                        BinaryContent binaryContent = new BinaryContent(
+                                file.getOriginalFilename(),
+                                (long) file.getBytes().length,
+                                file.getContentType(),
+                                file.getBytes()
+                        );
+                        binaryContentRepository.createBinaryContent(binaryContent);
+                        binaryContentIdList.add(binaryContent.getId());
+
+                    } catch (IOException e) {
+                        throw new FileException(e.getMessage());
+                    }
                 }
             }
         }
@@ -72,7 +85,7 @@ public class BasicMessageService implements MessageService {
     }
 
     @Override
-    public MessageUpdateResponse updateMessage(MessageUpdateRequest request) {
+    public MessageUpdateResponse updateMessage(MessageUpdateRequest request, List<MultipartFile> files) {
         //입력값 검증 처리하겠습니다
 //        validateUUID(request.messageId());
 //        validateString(request.content());
@@ -81,22 +94,36 @@ public class BasicMessageService implements MessageService {
         Message messageTemp = messageRepository.findMessageById(request.messageId())
                 .orElseThrow(() -> new ObjectNotFoundException("에러: 해당 메시지는 데이터파일에 존재하지 않습니다."));
 
-        //이전 첨부파일 삭제
-        for (UUID attachmentId : messageTemp.getAttachmentIds()) {
-            binaryContentRepository.deleteBinaryContent(attachmentId);
-        }
-
-        //첨부파일 추가 작업
-        List<UUID> binaryContentIdList = new ArrayList<>();
-        if (request.attachmentPathList() != null) {
-            for (String attachmentPath : request.attachmentPathList()) {
-                if (attachmentPath != null && !attachmentPath.isBlank()) {
-                    //binaryContent 생성
-                    BinaryContent binaryContent = new BinaryContent(attachmentPath);
-                    binaryContentRepository.createBinaryContent(binaryContent);
-                    binaryContentIdList.add(binaryContent.getId());
+        List<UUID> binaryContentIdList = messageTemp.getAttachmentIds();
+        //첨부파일 존재 시
+        if (files != null) {
+            //이전 첨부파일 삭제
+            if (binaryContentIdList != null) {
+                for (UUID attachmentId : binaryContentIdList) {
+                    binaryContentRepository.deleteBinaryContent(attachmentId);
                 }
             }
+
+            List<UUID> newBinaryContentIdList = new ArrayList<>();
+            for (MultipartFile file : files) {
+                if (file != null && !file.isEmpty()) {
+                    try {
+                        //binaryContent 생성
+                        BinaryContent binaryContent = new BinaryContent(
+                                file.getOriginalFilename(),
+                                (long) file.getBytes().length,
+                                file.getContentType(),
+                                file.getBytes()
+                        );
+                        binaryContentRepository.createBinaryContent(binaryContent);
+                        newBinaryContentIdList.add(binaryContent.getId());
+
+                    } catch (IOException e) {
+                        throw new FileException(e.getMessage());
+                    }
+                }
+            }
+            binaryContentIdList = newBinaryContentIdList;
         }
 
         log.info("메시지: {}가 수정됨.\n->{}", messageTemp.getContent(), request.content());
